mysql ox plugin push impl.

+++ plugins/mysql_ox_plugin/src/database/models/labels.rs
@@ -1,15 +1,20 @@
-use sqlx::{FromRow, Type};
+use ormx::Table;
+use sqlx::FromRow;
 
-#[derive(Debug, Clone, FromRow, Type)]
+#[derive(Debug, Clone, FromRow, Table)]
+#[ormx(table = "labels", id = id, insertable)]
 pub struct Label {
-    pub id: i32,
+    #[ormx(default)]
+    pub id: u64,
     pub name: String,
     pub value: String,
 }
 
-#[derive(Debug, Clone, FromRow)]
+#[derive(Debug, Clone, FromRow, Table)]
+#[ormx(table = "common_labels", id = id, insertable)]
 pub struct CommonLabel {
-    pub id: i32,
+    #[ormx(default)]
+    pub id: u64,
     pub name: String,
     pub value: String,
 }
